Resolve ENGKNOW-1182 "Int() should not throws an error when the column is a number in string format"

+++ gortools/src/main/scala/gorsat/parser/TypeConversionFunctions.scala
@@ -125,7 +125,7 @@ object TypeConversionFunctions {
 
   def string2Int(ex: sFun): iFun = {
     cvp => {
-      ex(cvp).toInt
+      ex(cvp).toDouble.toInt
     }
   }
